+++ tests/Js2IL.Test262.Tests/language/expressions/object/JavaScript/11.1.5_7-3-1.js
@@ -31,8 +31,7 @@ assert.compareArray = function(actual, expected) {
     }));
 };
 
-var o;
-  eval("o = {set foo(arg){return 1;}};");
+var o = {set foo(arg){return 1;}};
   var desc = Object.getOwnPropertyDescriptor(o,"foo");
 
 assert.sameValue(desc.enumerable, true, 'desc.enumerable');