+++ tests/Js2IL.Test262.Tests/language/expressions/object/JavaScript/11.1.5_7-3-2.js
@@ -31,8 +31,7 @@ assert.compareArray = function(actual, expected) {
     }));
 };
 
-var o;
-  eval("o = {set foo(arg){}};");
+var o = {set foo(arg){}};
   var desc = Object.getOwnPropertyDescriptor(o,"foo");
 
 assert.sameValue(desc.get, undefined, 'desc.get');